+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unpublished
 
+* **Breaking** `get_font_use` now gets font references, not just names.
 * Add `ZNG_NO_CRASH_HANDLER` env var to easily disable crash handler for special runs like for a debugger.
 * Add `CrashConfig::no_crash_handler` for custom crash handler disabling.
 * Add `zng::app::print_tracing_filter`.